Relocate project delete dialog
