+++ asreview/webapp/src/HomeComponents/DashboardComponents/ProjectTable.js
@@ -18,7 +18,8 @@ import {
 } from "@mui/material";
 import { styled } from "@mui/material/styles";
 
-import { ProjectDeleteDialog, TableRowButton } from "../DashboardComponents";
+import { ProjectDeleteDialog } from "../../Components";
+import { TableRowButton } from "../DashboardComponents";
 import { ProjectAPI } from "../../api/index.js";
 import { useRowsPerPage } from "../../hooks/SettingsHooks";
 import { useToggle } from "../../hooks/useToggle";
