+++ asreview/webapp/src/HomeComponents/DashboardComponents/index.js
@@ -1,5 +1,4 @@
 export { default as DashboardPage } from "./DashboardPage";
 export { default as NumberCard } from "./NumberCard";
-export { default as ProjectDeleteDialog } from "./ProjectDeleteDialog";
 export { default as ProjectTable } from "./ProjectTable";
 export { default as TableRowButton } from "./TableRowButton";
